updated the formatted date functions into a new file for clarity!
