+++ Achilles/ItemDisplayView.swift
@@ -348,19 +348,5 @@ struct VideoPlayerPlaceholderView: View {
     }
 }
 
-// MARK: - Date Suffix Helper
-func daySuffix(for date: Date) -> String {
-    let calendar = Calendar.current
-    let day = calendar.component(.day, from: date)
-    switch day {
-    case 11, 12, 13: return "th"
-    default:
-        switch day % 10 {
-        case 1: return "st"
-        case 2: return "nd"
-        case 3: return "rd"
-        default: return "th"
-        }
-    }
-}
+
